Migrate remaining utilities tests
